Remove advisory images-v2-import capability
It's premature to include an Images v2 API import capability in DefCore
(even in an advisory capacity) given that image import functionality has
yet to be developed in Glance.

The change proposed by this patch is in accord with the TC resolution
"Compute Requirements for Images and Kernels" [0], which states that "An
OpenStack Powered Compute Cloud MUST allow end user Image Uploads".
(End user image upload is typically referred to in the industry as
"image import" in order to distinguish it from from the simple upload
facility used by trusted sources, for example, an OpenStack service like
Nova, or Glance administrators supplying public images for use in a
cloud).  I interpret the TC resolution as a forward-looking statement
saying that the capability must be exposed as soon as it's available,
not that the capability must be made available immediately using the
(possibly improper) means currently at hand.  In other words, if we want
to impose such a requirement on OpenStack deployers, then it's incumbent
upon OpenStack to provide a means by which this can responsibly be done
in an interoperable way.  That means is scheduled to be delivered in
Glance in the Mitaka cycle.

The Glance image import spec [1] contains discussion and links to even
more discussion concerning this issue.  There's also lively discussion
of the issue on the gerrit review for the spec. [2]

The images-v2-import capability should be resinstated as an advisory
capability after the Mitaka release.  (The set of tests defining the
capability, however, will have to be modified.)

[0] https://governance.openstack.org/resolutions/20151211-bring-your-own-kernel.html
[1] http://specs.openstack.org/openstack/glance-specs/specs/mitaka/approved/image-import/image-import-refactor.html
[2] https://review.openstack.org/#/c/232371/

Change-Id: I98714d97a80a14b62b936ea51ccbde50117401b5