Remove test_reboot_server_soft
This test has been flagged for some time due to a longstanding
bug [1] in Tempest (open since 2012).  In discussion with QA, it appears
that the bug isn't going to be fixed and the test is likely to be
removed.  Since the tests has been flagged for several cycles now
and the test itself is likely to be dropped, we should remove it
from the 2016.01 Guideline and future Guidelines.

[1] https://bugs.launchpad.net/tempest/+bug/1014647

Change-Id: Ia1adc9e6b3f81c65922b5167d99347d47bc95ac7