Remove old 2016.next.json file
Last year we made a change [1] so that the in-development Guideline
file would ne named "next.json" instead of "$year.next.json".
At the time, we left the old 2016.next.json file in the tree, but
changed it's contents to be a simple pointer to the new location.
The file also contained a note that reads:

"Please remove this file after 2016 Jan."

This patch does just that.

[1] https://review.openstack.org/#/c/236729/

Change-Id: Ib3b367ef45753d17d701bcc4468d2e7d64a155c5