Add scoring for Neutron subnet pools
Following discussion with the Neturon PTL and other community members,
this patch adds scoring for the 2016.08 Guideline for Neutron.  In
the last Guideline we introduced quite a lot of Neutron capabilities
(there were none previously), so this was expected to be a relatively
small bump.  In the end, no new Capabilities emerged that warranted
inclusion.  The Subnet Pools feature implemented starting in the Kilo
release is generally regarded as something that will likely become
a Capability at some point down the road, but after some research it
doesn't appear to meet enough Criteria for inclusion today.

Although no new Capabilities are being introduced as advisory
in the JSON file, this patch adds a research note to the scoring sheet
concerning subnet pools in order to document some of the research done
during this cycle and help make future scoring of the capability easier.

Change-Id: I282e9b7c23a413da76370923037cb26ac576a65f