Move advisory -> required in .next
As discussed at the DefCore spring 2016 midcycle [1], we feel that most
or all of the Capabilities listed as advisory in the 2016.01
Guideline are now ready to move to required status.  This patch
moves them into required status starting with the 2016.08 Guideline,
and removes a few Capabilities that were deprecated in 2016.01.

An extraneous heading for volumes-v2-create-from-snapshot has been
removed as this grouping was dropped from the final version of the
Cinder scoring in 2016.01 [2] but we apparently forgot to remove
the header.

[1] https://etherpad.openstack.org/p/DefCoreSpring2016MidCycle
[2] https://review.openstack.org/#/c/221631/14/2016.next.json

Change-Id: Id3a302099fb5efd0e80bdb576537f643948713ea