Remove test that requires admin credentials
In the initial Neutron scoring for routers, we added a test that
requires admin credentials: test_create_router_setting_tennant_id.
This patch removes the test since DefCore typically does not
use tests that require admin access as interoperability tests.

The test in question is currently in advisory status and has not
yet become required, so it can be removed without a deprecation
cycle.

Change-Id: I5303b9fa2afa5e0388c442aa433891eaac93c89a