Added keystone scoring for Mitaka release cycle.
Scored an existing advisory capability (identity-v3-api-discovery) and
made it required in next.json.

Scored 3 new capabilities for which I couldn't find tests, adding
them however for discussion. One of the new capabilities,
identity-v3-catalog refers to the catalog that is returned when calling
the current required capability identity-v3-tokens-create. Users may be
relying on the catalog to be there, so I think it is worth discussing.

identity-v3-list-projects and identity-v3-list-groups are worth
discussing and consider for addition of new tests.

The rationale for this scoring as well as input from keystone's PTL can
be found in working_materials/keystone_capabilities_info.csv.

Change-Id: Id444f5e982f2e81f140e285c305e9c322f5b9f42