Leave capabilities advisory until tests updated
In the process of creating the 2016.08 Guidelines, we identified
several Capabilities that met Criteria to be added as required,
and had been identified as Advisory in the previous Guideline.
Subsequently, the networks-l3-router, networks-l3-CRUD, volumes-v2-qos,
and volumes-v2-transfer Capabilities were moved to required status
(as part of a larger move of 25+ Capabilities to required).  However,
we later discovered[1] that the tests for these Capabilities require
admin credentials even though the API's under test typically do not.

Since DefCore tests may not use admin credentials, these tests
cannot be made required.  Since all tests for these Capabilities
require admin credentials, the Capabilities would be left with no tests
and therefore cannot become required either.  The DefCore Committee
is exploring the possibility of finding alternate tests for these
Capabilities or modifying the existing tests to not require admin
credentials (since, as noted, the requisite API's themselves are
in fact typically exposed to and used by tenants).  In the interim,
we will leave the Capabilities in advisory status per discussion
at the 2016-04-06 DefCore Committee meeting [2] in order to reduce
confusion.

[1] https://review.openstack.org/#/c/299491/
[2] http://eavesdrop.openstack.org/meetings/defcore/2016/defcore.2016-04-06-16.01.html

Change-Id: I9af4bda7dafe54aed3cadb0b4ea32c1e266ee697