Add cutoff_score to next.json, update script
We recently added a field in Schema 1.4 to record a
"cutoff_score".[1]  The field reflects the minimum score that a
Capability needs to achieve for a given Guideline in order to
warrant inclusion.  The current draft Guideline, however, hasn't yet
 been updated to include a cutoff score.

This patch adds a cutoff_score field to next.json and sets it to
a preliminary vale of 74 (which is indicative of the score we've
used in the past as a cutoff, but which can be changed by the
DefCore Committee any time between now and when the Guideline goes to
the Board for approval before it becomes binding).  The patch also
updates the tabulate_scores.py script to read the cutoff_score from the
JSON document and adds an asterisk (*) to lines in the scoring file whose
totals meet or exceed the cutoff as a visual indicator of which
Capabilities have scored highly enough to warrant inclusion in the
Guideline.  Verbage in the header of the scoring.txt file has also been
updated to reflect the meaning of the indicator.

The patch also adds the cutoff_score field to the 2016.01 document,
where it was inadvertently left off since the field was only added
to the schema in close proximity to the document being finalized for
Board approval.

[1] https://review.openstack.org/#/c/224868/

Change-Id: I73d08baae13f373c955ff94640fac43109125886