Close coverage gap after new OpenStack releases
Currently, each DefCore Guideline covers the most recent three
releases of the OpenStack Software at the time of the Guideline's
approval (which is generally 3-4 months after the release of a
new version of the software).  This timing means that there's a
3-4 month period during which products built on the most recent
OpenStack release may not be able to apply for an OpenStack
Powered(TM) license agreement because the version of OpenStack on
which they are built isn't covered by a Board-approved Guideline.

Whereas the OpenStack community and Foundation have an expressed
interest in enabling users to adopt new software quickly, this
patch closes the 3-4 month window by amending the 2016.01
Guideline (and future Guidelines) to allow it to cover the most
recent version of the software starting on the day that the
software is released.

Note that because this change entails a modification to a Board-
approved Guideline document, this change should not merge until
after the Board of Directors has approved the change.  The
matter will be discussed during the 2016-04-24 Board meeting.

Change-Id: I6db11762e48ff3297eb9f90c93fb68717b0f3d43