Fix Criteria defs in next.json
Last year we adjusted the definitions of the tools and clients Criteria
to make them focus on usage by software external to the OpenStack
community and internal to the OpenStack community respectively [1].
However, the definitions in the next.json document still reflect the
old, outdated definitions.  This patch sets that to rights by adjusting
the definitions in the next.json document to reflect the current state
of affairs.

Note: I considered just removing the definitions here entirely, but
it seems logical to keep them since the definitions in the CoreCriteria
document can and do change over time (so it's nice to know how the
Criteria were defined at the time the Guideline you're looking at
was approved).

[1] https://review.openstack.org/#/c/226978/

Change-Id: I57af817659a0d66a9d6c3c0d33e92e329033d5b5