Remove components in both advisory and required
In https://review.openstack.org/#/c/290271 the images components
were added to the advisory list but they were already present in
the required list. As far as I can tell this duplication was
accidental. This patches removes all the duplicated components
from the advisory list except one.

Images-v2-import was removed from the required list and left in
the advisory list instead. That component was not in the previous
2016.01 guideline. New items need to spend one at least one
cycle as advisory before they can become required.

Change-Id: I062015228ace55e71abd63b82c6a07c6a86506a5