Flag Compute tests which depend on the os-networks extension
Certain Compute tests [1] require the fixed_network_name
configuration to be set or for a provider network to be
defined. Setting either of these triggers a code path
that uses the os-networks extension to look up the
network's id. If this extension is not enabled, then
any test that creates a server will fail. The credential
provider needs additional logic to handle the case
where this extension is not enabled.

[1] https://github.com/openstack/tempest/blob/master/tempest/api/compute/servers/test_list_server_filters.py#L270

Change-Id: I34b8cc50f35d7207d12c73f980966066eddbcf81
Related-Bug: #1592173
Related-Bug: #1600349