Add alias for rebuild_deleted_sever test
The test_rebuild_reboot_deleted_server test was split into two
tests in December of last year:

https://review.openstack.org/#/c/239710/

We added an alias for this in the 2016.01 and 2015.07 Guidelines:

https://review.openstack.org/#/c/259226/

However it is not aliased correctly in the next.json file.  This
patch fixes the problem by adding an alias.  We should also
consider adding the other test (rebooting a deleted server), but
that should be done in a seperate patch.

Change-Id: Id280c8ea5a9eca5a384a4e5bdd90e2f5986538c1