Flag/drop snapshot_create_with_volume_in_use test
In 2016.01 we added a new advisory capability:
volumes-v2-snapshot-create-delete.  One of the tests for that Capability
involves creating a volume, attaching it to an instance, and then
creating a snapshot of the volume while it's in-use.  This is
potentially problematic for two reasons: first, it requires the use of
the force flag, which wasn't introduced until Liberty[1].  Kilo is still
covered until 2016.01 and 2016.08.  Second, snapshotting an in-use
volume is an operation which may produce corrupted data (which isn't
always well understood [2]) and therefore seems like a thing that some
clouds may not wish to offer to their clients.

Since the force flag wasn't available in all the releases covered by
the 2016.01 and 2016.08 Guidelines, we should probably wait to include
this test until a later release.  Waiting would also give us some time
to see if it's something clouds do indeed expose to their users...my
guess is that it is, but I could see why some clouds might not.

[1]
http://docs.openstack.org/admin-guide/blockstorage_volume_backups.html
[2]
https://support.rackspace.com/white-paper/allowing-snapshots-without-detaching-the-volumes/

Change-Id: I9dab755bde2f169d52503d47fcfca404b8777f77