Flag/drop test_create_update_port_with_second_ip
In 2016.01 we introduced a new advisory test that involves creating
a Neutron port with a fixed IP address and then updating it to add
a second IP address to the same port.  However, the nunber of fixed
IP addresses that can be assigned to a port is dependent on the
plugin that Neutron is using--and not all plugins support more than
one IP address per port (the maximum number of IP's per port is
configurable in neutron.conf and clouds whose drivers do not support
multiple fixed addresses per port simply set it to 1).  Further,
some clouds limit the number of IP's per port for security reasons
to prevent malicious activity (e.g. spoofguard).

This patch adds an informational flag to the 2016.01 guideline
(where the test is advisory) and removes it from next.json.

Change-Id: Ie7d84d5d608bfc86501a5f7e097373c4565e4eb0