Create first Interop Issues Report
The DefCore Committee has been working to provide a periodic report [1]
on a tractable, small number of major barriers we have uncovered
during the course of work to the community that we wish to focus
attention on.  The report is intended to be constructive rather than
critical, and assist the community in recognizing barriers to
interoperability so that they can be addressed.

This commit represents the first such report to be created.  The issues
included in this report were finalized at the DefCore Committee's summer
2016 sprint.[2]

[1] https://etherpad.openstack.org/p/DefCoreInteropReport

Change-Id: I01667ecb46f3fc0cacfccbd4be66a0249b9f60c7