Approve 2016.08, supersede 2015.07
This commit changes the 2015.07 Guideline from "approved"
to "superseded" and changes the 2016.08 Guideline from "draft"
to "approved" status.  This patch should not be merged until
the Board of Directors has approved the 2016.08 Guideline.  Voting
is planned for the August 23 Board meeting.

Note: this change also regenerates documentation for all other
superceded Guidelines.

Change-Id: I630978bdff9c166a4a3340f8c26bcdcbf7013cbf