Prepare next.json for 2017.01 scoring
This patch does two basic things:

1.) It cleans up a few outstanding differences between next.json
and 2016.08.json.  These were minor schema updates and a few
things that got changed in one place but not the other by accident
late in the scoring cylce.

2.) It adjusts the releases covered and target approval date for
the next Guideline.  Please note that the date given here is
an estimate based on historical dates...the date of the January
2017 Board meeting hasn't yet been finalized/posted.

With this patch in place, next.json should be ready to start
receiving proposals for additions to the next Guideline as scoring
activities for 2017.01 commence.

Change-Id: Ibb05fb6f7e393a5440ce34c177cf167a10d26311