Add info on Powered adoption
In the last Board of Directors meeting, one piece of feedback we
received on the new periodic issues report as a suggestion that we
include some information about what products have adopted the
Interoperability Guidelines and been granted an OpenStack Powered
agreement by the Foundation.  This patch adds some stats with links to
the Marketplace (which will contain current info long after the report
is published, as opposed to a static list of names in the report itself
which will become inaccurate over time).

Change-Id: I941f8c759c10d2d45911eec7f5d6fe11db57f234