2017.01 Scoring for Neutron
This change constitutes a first pass at scoring Neutron capabilities
for the 2017.01 Guideline.  It adds two new capabilities: subnet pools
and API version listing.  It also adds a new designated section:
neutron-lib, which is imported in several already-required capabilities.
It holds off on moving some L3 capabilities from advisory to required
for the moment as currently the tests still (unnecessarily) require
admin credentials, though patches are in progress to remedy that situation.

Change-Id: Ic684f170209dac64f65cba3ab460eb917d781cd4