Add Guidance on Modifying Tests, Filing Flags
It's now come up a couple of times that we've found people modifying
tests in order to pass them.  Although seems obvious that not testing
things the same way across all clouds voids any guarantee of
interoperability, we should make this explit in the guidance we provide.
It's been proposed that one reason folks may be modifying tests is that
they don't clearly understand what they need to do in order to file a
flag request when they run into a problem, so they try to fix what they
think is wrong with the test instead.

This patch makes three changes to help address the issue:

1.) It adds a section to the 2016.01 guidance (which is currently
linked to from http://openstack.org/interop) to make it explicitly
clear that tests shouldn't be modified.

2.) Assuming that the interop website will be updated to point to the
most recent Guideline at some point, it creates the same guidance for
2016.08.

3.) It ammends HACKING to include simple instructions and an example
of how to file a flag request.  The individual guidelines' guidance
files also point to HACKING so folks know to look there.

Change-Id: I2f57a852da3181714e87d8e689dd5a1cb33cb417