Fixed malformed 1.6.json schema
The 1.6.json schema included an error where an additionalProperties
tag was erroniously attached to a properties object, rather than
an object definition. A recent update to jsonschema improved
checking and flagged this previously uncaught error. This patch
fixes the malformed schema, and does not materially change the
content or intent of the original file.

Change-Id: I42c35b65aa622a2ea7205a728e8431b4f0e6daf6
Closes-Bug: #1664113