Fix incorrect test case name (fully qualified test name)
Test [1] does not included class name in its FQN.  Thus, no Tempest
version will be able to run this test because its name won't be able
to match any name from the test list produced by Tempest.  Fix
2017.01.json by adding the correct name as alias since this is an
approved guideline. For next.json, the incorrect name is replaced
with the correct one because it does not make sense to add a
non-existing test name as alias.

[1] tempest.api.identity.v3.test_list_projects_returns_only_authorized_projects

Change-Id: Ice75be2402d65ff473f68aa77db66167ae452363