Deprecate volume-v2/v3-attach-detach capability
The rationale for this patch has been stated in:
http://lists.openstack.org/pipermail/interop-wg/2017-April/000116.html

In short, the os-attach/detach in Cinder only changes the volume
state, it does not actually attach/detach the volumes, which is already
covered by "compute-volume-attach" capability. Therefore there is no
need for "volume-v2-attach-detach" capability of which test could break
defcore testing for OpenStack clouds, especially for public clouds.

Based upon the above reasoning, this patch accomplish two things:
(1) Flag all the test cases under volume-v2-attach-detach in
2016.08.json and 2017.01.json, and all the test cases under
volumes-v3-attach-detach in 2017.01.json.
(2) Move volume-v2/v3-attach-detach from "required" to "deprecated"
section in both 2017.01.json and next.json, and remove the detailed
test sections below.

Action in (1) is mainly due to the fact that defcore does not have
a mechanism to flag capability. Therefore the only way I could do
is to flag all the related tests with [D400] which indicates that
the capability associated with the tests has problems.

Change-Id: Iabb1d0820300442e407d8b92dfb575ac71471d06
Signed-off-by: zhipengh <huangzhipeng@huawei.com>