Add Aliases For VolumeV2 Test Cases Part 1
The recent version of Tempest (at least from v15) has renamed
all the previous VolumeV2* test case to Volume*. However the
current defcore guideline has not added such alias, which has
resulted in error when user wants to use new version of tempest
to test.

This patch is the part 1 of the fix, due to the large amount of
work. This patch has provided necessary aliases for all the
volume-v2-* capabilities (that means mission complete for required
capabilities) and some of the volume-v3-* ones.

We should expect another patch to provide the aliases for the
remaining volume-v3-* which are at the moment advisory capabilities

Co-Authored-By: Luz Cazares <luz.cazares@intel.com>
Signed-off-by: zhipengh <huangzhipeng@huawei.com>

Change-Id: I700b7009bfb3a93ade832198f4d58f487d641af8