Add Aliases For VolumeV2 Test Cases Part 2
This is the second patch for I700b7009bfb3a93ade832198f4d58f487d641af8
It provide all the aliases for volume-v3-* which are
advisory capabilities.

The change is due to tempest renaming of its VolumeV2 classes.
Tempest change is to reuse the same test case for both V2 and V3.
Avoiding confusion about the version being tested by removing "V2"
from the name. Notice that Cinder version should be setup at
tempest.conf file and defaults to V2.

Change-Id: I5d709698297d043b5cb4a40c263f30cc00908238