glance scoring for the 2018.01 guideline
There are a few loose ends on this one, to tie up, but
this is my first draft. There have not been a great number
of changes to Glance in the last cycle. The greatest change
that may be in the works is the finalization of the images-v2-import
capability. I will be checking up on whether or not that has, as of
yet, been finalized.
Change-Id: Ie318ac37833dd4836fd08b27e7ec738694a399db