First attempt at scoring trunk ports for NFV vertical
This is the first attempt at scoring the Neutron trunk port
feature as initial capability for the new NFV vertical program.
Moreover, a first version of a NFV guideline based on the 2.0
schema definition is included.

Since the NFV program is a vertical, I scored all 12 criteria -
in contrast to the add-on programs which are limited to 9
criteria.

I scored the 'widely deployed' criterion' with 1 in the context
of NFV-focused deployments. Since those are mostly private
cloud deployments, it is hard to determine actual availability.

Change-Id: I496ff4abb9c2510f0676f2f849414be785fb162b