Scoring of networking-bgpvpn project for NFV vertical
Scoring, notes and guideline updates for the networking-bgpvpn project.
From a scoring perspective, this looks fine for inclusion in the NFV
vertical. One of the capabilities is an admin-only API but this
criterion might be relaxed for the NFV vertical. The test coverage needs
to be discussed and potentially extended.

Change-Id: I415bcd3d5bc66708c12af6c8988d1f4f101c51e1