Create 2018.02 document for approval
This commit creates the 2018.02 guideline that will be presented to
the Board of Directors at the February 2018 meeting for approval.
This commit moves all capabilities that were advisory in 2017.09
to required status.

This patch also refreshes the next.json document to prepare for the
fall 2018 guideline drafting process.  As the meeting date for the
BoD meeting at the fall PTG has not yet been determined, it uses an
arbitrary date in August for the time being (we can change this
later once we have a firm date).  It also moves items that were
advisory in 2018.02 to required status.

This patch can be merged now as it sets the 2018.02 document to
advisory status--we will change it to approved if and when the BoD
votes to approve it next month.

Change-Id: I634e2dc1273e7b67103c86436b4ea5489cae3543