Move next.json to schema 2.0
The 2018.02 Guideline was the first to be published in the new
schema 2.0 format. [1]  The change to 2.0 was done with the agreement
that it would probably be easiest to write 2018.02 in the new format
and then create a new next.json from it (rather than the other way
around) due to the timing.  This patch transforms next.json to
schema 2.0 by using the existing 2018.02 document as the base.  Once
this lands, we'll be able to begin work on the next round of scoring
using the new 2.0 format, as well as do the usual round of housekeeping
like removing old capabilities and moving advisory ones to required.
For simplicity, we'll do those in future patches.

This patch also removes one test from the listing as it has been
combined with another test under the other test's idempotent ID.
See discussion in [2] for more background.  Leaving the test here
but also swapping the name and alias as we normally do in next.json
would cause a duplicate key issue with the docs jobs.

[1] https://git.openstack.org/cgit/openstack/interop/schema/2.0.json
[2] https://review.openstack.org/#/c/548990/3/2018.02.json@2924

Change-Id: I12457d66e804f6f6673c604542848d41ac1396a2