Clean up advisory list
In 2018.02 we set the "required_since" field to "2018.02" on a couple
of capabilities to mark them as moving from advisory to required.
However, they're still in the "advisory" list in the components section.
This patch cleans them up by removing them from the advsiory list and
placing them in the required list.

We also assigned volumes-v3-upload a high enough score to qualify as
advisory and added it to the JSON document, but did not add it to the
advisory section.  This patch adds it.

Further, this patch removes two capabilities from the "removed" list
as is customary after a cycle.

Change-Id: Ib7c1216f3b3045e709e7fefa8b7ccce9922959b5