Add compute-servers-create-multiple as advisory
This patch adds a new capability for Nova: the "create multiple
servers" API [1], which allows a user to create multiple servers with
a single API call.  Although this API is not widely used by tools such
as gophercloud, libcloud, or Terraform, it meets most other criteria
and is not specific to any particular backend.

[1]
https://developer.openstack.org/api-ref/compute/#create-multiple-servers

Change-Id: I5bdbc1504fb088392e55b76c2ce70f19d6aa69c6