Remove Cinder deprecated Capabilities
The Cinder v2 APIs are deprecated since Pike
- https://docs.openstack.org/releasenotes/cinder/pike.html#deprecation-notes
and its capabilities were also marked as Deprecated in
- Ie104503e1655e1dad4737b86bc047c6c363a6569.

Same for volume reserve capabilities which were marked
as Deprecated in
- I1599c1c3721e72a415be65af8b474206b174f241

Pike release was last supported in 2018.11 guidelines and
2019.11 guidelines onwards, Cinder v2 and reserve capabilities
were not included in guidelines.

This is time to move those Deprecated capabilities in the Removed list
and later in the next cycle we can completely cleanup those as well
as their test information.

Change-Id: Ic0d18d89b917d0fd428ecbde40c9a2dd7682eeee