Fix gate interop-tempest-consistency job
interop-tempest-consistency job still use the python2.7
env to run as default basepython in testenv is py2.7.

Requirement has dropped the py2.7 caps now
- I2d58d1b524764ea1d46d6ab3a382783f6a002b34

and interop-tempest-consistency job started failing on interop
and Tempest gate. To unblock gate on both project we need to move
the default tox env of interop to python3.

Change-Id: I40770cf254369808c2e420dae6c2ec52bf47b777